Remind command and Upload Ui Screenshot
