+++ src/main/java/bobi/exception/InvalidCommandException.java
@@ -1,0 +1,17 @@
+package bobi.exception;
+
+/**
+ * InvalidCommandException inherits from BobiException.
+ * An InvalidCommandException object encapsulates the error message
+ * which will be shown when this exception is thrown.
+ *
+ * @author ruo-x
+ */
+public class InvalidCommandException extends BobiException {
+    /**
+     * Constructor of an InvalidCommandException object.
+     */
+    public InvalidCommandException() {
+        super("Oh no! Bobi does not know what that means... :/");
+    }
+}
