+++ src/main/java/bobi/gui/Launcher.java
@@ -1,0 +1,14 @@
+package bobi.gui;
+
+import javafx.application.Application;
+
+/**
+ * Launcher class launches the GUI of Bobi.
+ *
+ * @author ruo-x
+ */
+public class Launcher {
+    public static void main(String[] args) {
+        Application.launch(Main.class, args);
+    }
+}
