--- src/main/java/duke/gui/Launcher.java
@@ -1,14 +1,0 @@
-package duke.gui;
-
-import javafx.application.Application;
-
-/**
- * Launcher class launches the GUI of Bobi.
- *
- * @author ruo-x
- */
-public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-    }
-}
